Safari-611.1.10/LayoutTests/editing/execCommand/4786404-2.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/dump-as-markup.js"></script>
<div id="container"><div id="div" contenteditable="true"><u>foo</u> bar baz</div></div>
<script>
var div = document.getElementById("div");
div.focus();

document.execCommand("SelectAll");
document.execCommand("RemoveFormat");

Markup.description('This tests for a bug where performing the Remove Format operation on a selection that started in underlined content would underline everything operated on. "foo bar baz" should not have underline.');
Markup.dump('container');

</script>
</body>
</html>
